Added for real simulation

+++ AutoDataGenerator/measurementRequest.cs
@@ -5,6 +5,6 @@ namespace GenerateRandomSensorData
     public class measurementRequest
     {
         public string Tag { get; set; }
-        public float Measurement { get; set; }
+        public double Measurement { get; set; }
     }
 }
